fix: Move path to domain
