+++ src/domain/BridgeGame.js
@@ -1,5 +1,5 @@
-const { GAME_CONDITION } = require("./utils/Constants");
-const BridgeGameResult = require("./BridgeGameResult");
+const { GAME_CONDITION } = require("../utils/Constants");
+const BridgeGameResult = require("../BridgeGameResult");
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
