+++ src/domain/BridgeGameResult.js
